requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.0
channels==4.3.2
channels_redis==4.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==4.2.27
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-db==0.0.7
django-debug-toolbar==6.1.0
django-filter==25.1
django-jazzmin==3.0.1
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
gunicorn==23.0.0
kombu==5.6.1
msgpack==1.1.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyMySQL==1.1.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.0.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unicorn==2.1.4
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
 